Add --to-code to chain encode params
